Fix ODO_STEPS data type from uint8_t to uint16_t to handle 4096 steps

+++ gauge_V4/config_calibration.cpp
@@ -43,7 +43,7 @@ unsigned int TACH_MAX = 6000;       // RPM at shift point
 unsigned int TACH_MIN = 3000;       // Minimum RPM to show
 
 // ===== ODOMETER MOTOR CALIBRATION =====
-uint8_t ODO_STEPS = 4096;             // Steps per revolution
+uint16_t ODO_STEPS = 4096;             // Steps per revolution
 uint8_t ODO_MOTOR_TEETH = 16;       // Number of teeth on motor gear
 uint8_t ODO_GEAR_TEETH = 20;        // Number of teeth on odometer gear
